% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base64.R
\name{b64enc}
\alias{b64dec}
\alias{b64enc}
\title{Encode to base64 or Decode from base64}
\usage{
b64enc(x)

b64dec(x)
}
\arguments{
\item{x}{[\code{character(1)}]\cr
Character vector to encode or decode.}
}
\value{
[\code{character}] of the same length as input \code{x}.
}
\description{
In contrast to RFC3548, the 62nd character ("+") is replaced by "-", the 63rd character ("/") is replaced by "_". 
Furthermore, the decoder does not insert "=" at the end of the string to pad to the right length.
As a result, all characters will comply to the pattern \dQuote{[A-Za-z0-9_-]} and can be used on basically all 
file systems and in URLs.
}

