% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base32.R
\name{base32_encode}
\alias{base32_encode}
\alias{base32_decode}
\title{Encode to base32 or Decode from base32}
\usage{
base32_encode(x, use.padding = TRUE)

base32_decode(x, use.padding = TRUE)
}
\arguments{
\item{x}{[\code{character(1)}]\cr
Character vector to encode or decode.}
}
\value{
[\code{character}] of the same length as input \code{x}.
}
\description{
Simple RFC4648 base32 encoder/decoder. Pads with \dQuote{=}.
}
\examples{
x = "plain text"
encoded = base32_encode(x)
decoded = base32_decode(encoded)
print(encoded)
print(decoded)
}
